Reorganize components into major and minor directories
