+++ src/components/major/category-bar.tsx
