+++ src/components/major/events.tsx
