+++ src/components/major/nav-bar.tsx
@@ -1,11 +1,11 @@
-import DirectButton from './DirectButton';
+import DirectButton from '../minor/DirectButton';
 let id: number = 5; //for when we have user auth in place
 
 export default function NavBar() {
   if (id) {
     return (
       <div className="flex justify-between border-b-3 input">
-        <DirectButton text={'LIGHT MODE'} route={'/userhome/music'} />
+        <DirectButton text={'LIGHT MODE'} route={'/userhome'} />
         <div className="flex justify-between">
           <DirectButton text={'HOMEPAGE'} route={'/userhome'} />
           <DirectButton text={'MY EVENTS'} route={'/savedevents'} />
@@ -16,9 +16,11 @@ export default function NavBar() {
   } else {
     return (
       <div className="flex justify-between border-b-5 input">
-        <DirectButton text={'LIGHT MODE'} route={'/userhome'} />
+        <DirectButton text={'LIGHT MODE'} route={'/'} />
         <DirectButton text={'LOGIN'} route={'/login'} />
       </div>
     );
   }
 }
+
+//note to self - data flow through front end, refactoring and file structure
