+++ src/components/minor/DirectButton.tsx
@@ -8,7 +8,10 @@ interface DirectButtonProps {
 export default function DirectButton({ text, route }: DirectButtonProps) {
   return (
     <>
-      <Link to={route} className="text-xl p-1 ml-2 mr-2 hover:bg-[#E1D2E5]">
+      <Link
+        to={route}
+        className="text-xl p-1 ml-2 mr-2 hover:bg-[rgba(225,210,229,0.1)]"
+      >
         {text}
       </Link>
     </>
